+++ education/number/posit/basic_operators.cpp
@@ -1,12 +1,12 @@
 // basic_operators.cpp : examples of the basic arithmetic operators using posits
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/number/posit/posit.hpp>
 
 // quick helper to report on a posit's specialness
-template<size_t nbits, size_t es>
+template<unsigned nbits, unsigned es>
 void checkSpecialCases(sw::universal::posit<nbits, es> p) {
 	std::cout << "posit is " << (p.iszero() ? "zero " : "non-zero ") << (p.ispos() ? "positive " : "negative ") << (p.isnar() ? "Not a Real" : "Its a Real") << std::endl;
 }
@@ -16,8 +16,8 @@ int main()
 try {
 	using namespace sw::universal;	// standard namespace for posits
 
-	const size_t nbits = 16;
-	const size_t es = 1;
+	const unsigned nbits = 16;
+	const unsigned es = 1;
 	posit<nbits, es> p1, p2, p3, p4, p5, p6;
 
 	/* constexpr */ double minpos = double(posit<nbits, es>(SpecificValue::minpos));
